requirements.txt
discord.py
python-dotenv
aiohttp
google-cloud-aiplatform
google-cloud-spanner
